Ensure that EC2 instances are not gone for long from ASG
This is partly to help reduce confusion when developers are trying to work
out how many nodes they have, but it's also because Ophan's Elasticsearch
Node Protection won't change the protection status of an EC2 instance if
it's not in the Auto-Scaling Group. For us on Ophan, this meant that
ES-Node-Rotation was failing to kill the old node even once it had stopped
it's Elasticsearch service.

The change in this commit ensures that we re-attach the 'old' node to the
ASG once we're sure (thanks to `ClusterSizeCheck`) that the ASG has
brought up a 'new' node.

https://github.com/guardian/ophan/issues/2745

Co-authored-by: Amy Hughes <amy.hughes@guardian.co.uk>