Do not wait for irrelevant shard reallocations to complete
I believe non-Master node (eg a Data Node) can be killed if:

* it has no Documents
* [slightly paranoid] it is excluded from shard allocation

We know that the 2nd condition is already true as we've executed the
`MigrateShards` step. If the 1st condition is true, we should proceed and
kill the box - it **doesn't matter** if other shard reallocations are
taking place.

On our Ophan cluster, we find that shard reallocations can continue for
hours after the old cluster is completely empty, so waiting for _no_
reallocations to be taking place can actually time-out the ENP
step-function.

This fixes issue #31:

https://github.com/guardian/elasticsearch-node-rotation/issues/31