Ensure delay/failure of shutdown doesn't prevent termination of instance
We've seen a couple of times that the Step Function has not terminated
the old instance, due to a sluggish response on the systemd shutdown:

https://eu-west-1.console.aws.amazon.com/cloudwatch/home?region=eu-west-1#logEventViewer:group=/aws/lambda/enr-remove-node-PROD;stream=2019/01/24/[$LATEST]79dbdedef6b74c98849a0439a9501b8e;start=2019-01-23T15:24:09Z

```
15:24:09
2019-01-24T15:24:09.347Z	c69e513f-adb4-4f85-94c8-0abd59719ebb
Failed due to terminate i-0f88ebbef0f4b7bc8 due to: SSM command result was: InProgress /

15:24:09
2019-01-24T15:24:09.349Z	c69e513f-adb4-4f85-94c8-0abd59719ebb
{
    "errorMessage": "Failed due to terminate i-0f88ebbef0f4b7bc8 due to: SSM command result was: InProgress / "
}
```

Note that the systemd shutdown didn't even fail, it's just 'in progress'.
My belief is that the systemd shutdown is a nice-to-have if we already
know that the node contains no data - for an empty data node we could just
kill it already, for an elected master node it *is* good to attempt a
shutdown as this appears to improve the speed of master re-election - but
I'm not convinced that even then it's worth leaving the instance running.

Consequently, the change in this commit means that we only attempt a
systemd shutdown if the node is master-eligible - and even then, we don't
let it stop the termination of the instance.