Handle AWS API throttling in ASG scaling
This addresses https://github.com/guardian/elasticsearch-node-rotation/issues/45
Code is added to handle retrying failed AWS API calls, specifically
those that deal with attaching and detaching instances from ASGs.
We already have code to retry SSM commands but a different approach
is required when using the AWS API directly.

A `retry` function has been added that accepts a function (typically
anonymous) that will  produce the expected `Promise<T>` and that
can be called a  fixed number of times. After this `retry` will throw an
error that is handled by the AWS lambda runtime.

Co-authored-by: Roberto Tyley <roberto.tyley@guardian.co.uk>