Restrict SSM run command by Stack and Stage
At the moment the role can run any arbitrary commands on any instance in the account. It's not that big a deal since the role is only used within a lambda that is not publically accessible but if something were to exfiltrate those credentials it's totally game over for the entire account.

The tighest way of doing this would be to have a separate SSM document for each command permissioned accordingly, rather than allowing RunShellScript. That's quite heavyweight though and would be annoying for future developers trying to make small changes to each command.

Ideally we'd restrict even further by aws:autoscaling:groupName tied to the master, warm and hot autoscaling groups. That doesn't work unfortunately as the instance loses that tag once it is detached. In https://github.com/guardian/investigations-platform/pull/364 we use aws:cloudformation:stack-name but as ENR is a generic tool that already has Stack and Stage passed in, I decided to go with those for now. I haven't further restricted by App as we have various values for those across our many Elasticsearch clusters. It would be nice to use stack name or add App though, as even unrestricted shell access on a given Stack, Stage combo isn't great.