Retry Elasticsearch commands
We observed on a few occasions that node rotations failed
because requests to update cluster settings on the ES cluster
came back with `acknowledged: false`.

These updates are idempotent, so we would like to retry them,
rather than fail immediately.

More details here:
https://github.com/guardian/elasticsearch-node-rotation/issues/82

In addition to introducing retries, this also:

- Removes repetition between the implementation of the two cluster
  settings commands
- Introduces an elasticsearch client class that means the target
  node instance doesn't need to be passed around

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Object_initializer#computed_property_names