Make `app` consistent across all stacks Riff-Raff deploys to
The variation of `app` is a little confusing and doesn't appear to be needed.

In this change, we explicitly instruct Riff-Raff to discover the CFN stack to affect by providing the `cloudFormationStackName` parameter.

Removal of the `app` property should have the effect of:
  - RiffRaff finds a CFN stack via `cloudFormationStackName`
  - RiffRaff updates the `App` tag on the stack to match that of the [`template`](https://github.com/guardian/elasticsearch-node-rotation/blob/d6e07a4e511e16b12102c3e959b1430372382662/riff-raff.yaml#L8), i.e. making it consistent across instances of this stack
  - We can then, in a follow-up PR, move to discovering the CFN stack via tag lookup

Using a single, consistent, value makes migration to GuCDK easier.